+++ src/types/types.ts
@@ -1,4 +1,4 @@
-import { type ISdk } from "./core/sdk"
+import { type ISdk } from "./core"
 export type MaybePromise<T> = Promise<T> | T
 export type Maybe<T> = NonNullable<T> | undefined
 export type MaybeArray<T> = T | T[]